fix for project import stepper
