--- src/js/actions/MyProjectsActions.js
@@ -1,21 +1,0 @@
-import consts from './ActionTypes';
-// helpers
-import * as myProjectsHelpers from '../helpers/myProjectsHelpers';
-//helpers
-// contant declarations
-
-/**
- *  @description: With the list of project directories, generates an array of project detail objects
- */
-export function getMyProjects() {
-  return ((dispatch, getState) => {
-    myProjectsHelpers.migrateResourcesFolder();
-    const state = getState();
-    const { projectDetailsReducer: { projectSaveLocation } } = state;
-    let projects = myProjectsHelpers.getProjectsFromFS(projectSaveLocation);
-    dispatch({
-      type: consts.GET_MY_PROJECTS,
-      projects: projects
-    });
-  });
-}
