+++ src/js/containers/home/ProjectsManagementContainer.js
@@ -8,7 +8,7 @@ import ProjectsFAB from '../../components/home/projectsManagement/ProjectsFAB';
 import OnlineImportModal from '../../components/home/projectsManagement/OnlineImportModal';
 // actions
 import * as BodyUIActions from '../../actions/BodyUIActions';
-import * as MyProjectsActions from '../../actions/MyProjectsActions';
+import * as MyProjectsActions from '../../actions/MyProjects/MyProjectsActions';
 import * as ImportOnlineActions from '../../actions/ImportOnlineActions';
 import * as ImportOnlineSearchActions from '../../actions/ImportOnlineSearchActions';
 import * as OnlineImportWorkflowActions from '../../actions/Import/OnlineImportWorkflowActions';
